+++ src/store/providers/header-provider.tsx
@@ -2,7 +2,7 @@
 
 import { FC, ReactNode, useCallback, useEffect, useState } from 'react'
 import useMediaQuery from '@/hooks/use-media-query'
-import HeaderContext from './header-context'
+import HeaderContext from '@/store/header-context'
 import { MOBILE_BREAKPOINT } from '@/utils'
 
 type HeaderProviderProps = {
